+++ src/common/utils/deployUtils.ts
@@ -15,9 +15,9 @@ export async function forceSourcePush(scratchOrgAlias:string,debug = false) {
     await execCommand(pushCommand, this, { fail: true, output: true, debug: debug });
   } catch (e) {
     const {tips} = analyzeDeployErrorLogs(e.stdout + e.stderr);
-    uxLog(this,c.red("Sadly there has been Deployment error(s)"));
+    uxLog(this,c.red("Sadly there has been push error(s)"));
     uxLog(this,c.yellow(tips.map((tip:any) => c.bold(tip.label)+'\n'+tip.tip).join("\n\n")));
-    uxLog(this,c.yellow(c.bold(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the deployment issues :)`)));
+    uxLog(this,c.yellow(c.bold(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the push issues :)`)));
     throw new SfdxError('Deployment failure. Check messages above');
   }
 }
@@ -28,10 +28,10 @@ export async function forceSourcePull(scratchOrgAlias:string,debug = false) {
     await execCommand(pushCommand, this, { fail: true, output: true, debug: debug });
   } catch (e) {
     const {tips} = analyzeDeployErrorLogs(e.stdout + e.stderr);
-    uxLog(this,c.red("Sadly there has been Deployment error(s)"));
+    uxLog(this,c.red("Sadly there has been pull error(s)"));
     uxLog(this,c.yellow(tips.map((tip:any) => c.bold(tip.label)+'\n'+tip.tip).join("\n\n")));
-    uxLog(this,c.yellow(c.bold(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the deployment issues :)`)));
-    throw new SfdxError('Deployment failure. Check messages above');
+    uxLog(this,c.yellow(c.bold(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the pull issues :)`)));
+    throw new SfdxError('Pull failure. Check messages above');
   }
 }