+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image 
 
 - New command **hardis:source:push**
 - New command **hardis:source:pull**
+- Various mini-fixes
 
 ## [2.4.0] 2021-03-27